* master: (85 commits)
  Oops, forgot to bump the version in the CHANGELOG
  Fix Jeweler's Rubyforge settings for uploading docs
  Release 1.2.6
  Update CHANGELOG
  Add FakeWeb::VERSION so you can programmatically determine what version of FakeWeb is loaded without rubygems, etc.
  Switch to Jeweler
  Work around a ruby quirk where the literal /regexp/ syntax triggers "ambiguous first argument; put parentheses or even spaces" warnings
  Fix warnings like "method redefined; discarding old setup" in test_helper.rb
  Avoid double-requiring test_helper.rb when running rake tasks
  Enable warnings when running tests
  Add a little more request matching magic: registering a regex with an explicit port (like :get, /example.com:80/) now matches a request that uses an implied default port (like GET "http://example.com/").
  Restore that re-registering a URI replaces the first registration even when the query params are ordered differently
  Add some test coverage to ensure that re-registering a URI replaces the first registration
  Start refactoring the Registry to unify handling of Regexps and URIs/strings
  Fix that you'd have to sort the query params alphabetically in a registered regex in order for it to ever match a URI. Thanks to Ben Hall for discovering this.
  Improve some regex tests that weren't testing much
  Add missing coverage for FakeWeb.clean_registry
  Ignore Rubymine meta-info
  Add a test for some accidental functionality: you can register response headers using string option keys, too
  Release 1.2.5
  ...

Conflicts:
	.manifest
	fakeweb.gemspec
	lib/fake_web.rb
	lib/fake_web/registry.rb
	lib/fake_web/responder.rb
	test/test_fake_web.rb
	test/test_fake_web_open_uri.rb

+++ lib/fake_web/ext/net_http.rb
@@ -15,7 +15,7 @@ module Net  #:nodoc: all
       when Socket, OpenSSL::SSL::SSLSocket, IO
         io
       when String
-        if !io.include?("\0") && File.exists?(io)
+        if !io.include?("\0") && File.exists?(io) && !File.directory?(io)
           File.open(io, "r")
         else
           StringIO.new(io)
@@ -41,7 +41,8 @@ module Net  #:nodoc: all
       path = URI.parse(request.path).request_uri if request.path =~ /^http/
 
       if request["authorization"] =~ /^Basic /
-        userinfo = request["authorization"].sub(/^Basic /, "").unpack("m").first + "@"
+        userinfo = FakeWeb::Utility.decode_userinfo_from_header(request["authorization"])
+        userinfo = FakeWeb::Utility.encode_unsafe_chars_in_userinfo(userinfo) + "@"
       else
         userinfo = ""
       end
@@ -52,11 +53,12 @@ module Net  #:nodoc: all
       if FakeWeb.registered_uri?(method, uri)
         @socket = Net::HTTP.socket_type.new
         request.body = body if body
-        FakeWeb.response_for(request, method, uri, &block)
+        FakeWeb.response_for(method, uri, &block)
       elsif FakeWeb.allow_net_connect?
         connect_without_fakeweb
         request_without_fakeweb(request, body, &block)
       else
+        uri = FakeWeb::Utility.strip_default_port_from_uri(uri)
         raise FakeWeb::NetConnectNotAllowedError,
               "Real HTTP connections are disabled. Unregistered request: #{request.method} #{uri}"
       end
